- Remove unused cfg parameters in _check_type methods. - Move logging related code to _common.py module. - Change args prefix check to use startswith for clarity. - Simplify set_defaults implementation. - Move namespace_to_dict and dict_to_namespace to avoid string type hint.
