+++ jsonargparse/_namespace.py
@@ -94,28 +94,6 @@ def patch_namespace():
         argparse.Namespace = namespace_class
 
 
-def namespace_to_dict(namespace: "Namespace") -> Dict[str, Any]:
-    """Returns a copy of a nested namespace converted into a nested dictionary."""
-    return namespace.clone().as_dict()
-
-
-def dict_to_namespace(cfg_dict: Union[Dict[str, Any], "Namespace"]) -> "Namespace":
-    """Converts a nested dictionary into a nested namespace."""
-    cfg_dict = recreate_branches(cfg_dict)
-
-    def expand_dict(cfg):
-        for k, v in cfg.items():
-            if isinstance(v, dict) and all(isinstance(k, str) for k in v.keys()):
-                cfg[k] = expand_dict(v)
-            elif isinstance(v, list):
-                for nn, vv in enumerate(v):
-                    if isinstance(vv, dict) and all(isinstance(k, str) for k in vv.keys()):
-                        cfg[k][nn] = expand_dict(vv)
-        return Namespace(**cfg)
-
-    return expand_dict(cfg_dict)
-
-
 class Namespace(argparse.Namespace):
     """Extension of argparse's Namespace to support nesting and subscript access."""
 
@@ -352,6 +330,28 @@ def del_clash_mark(key: str) -> str:
     if key[0] == clash_mark:
         key = key[1:]
     return key
+
+
+def namespace_to_dict(namespace: Namespace) -> Dict[str, Any]:
+    """Returns a copy of a nested namespace converted into a nested dictionary."""
+    return namespace.clone().as_dict()
+
+
+def expand_dict(cfg):
+    for k, v in cfg.items():
+        if isinstance(v, dict) and all(isinstance(k, str) for k in v.keys()):
+            cfg[k] = expand_dict(v)
+        elif isinstance(v, list):
+            for nn, vv in enumerate(v):
+                if isinstance(vv, dict) and all(isinstance(k, str) for k in vv.keys()):
+                    cfg[k][nn] = expand_dict(vv)
+    return Namespace(**cfg)
+
+
+def dict_to_namespace(cfg_dict: Union[Dict[str, Any], Namespace]) -> Namespace:
+    """Converts a nested dictionary into a nested namespace."""
+    cfg_dict = recreate_branches(cfg_dict)
+    return expand_dict(cfg_dict)
 
 
 # Temporal to provide backward compatibility in pytorch-lightning
