+++ jsonargparse/_signatures.py
@@ -8,7 +8,7 @@ from contextlib import suppress
 from typing import Any, Callable, List, Optional, Set, Tuple, Type, Union
 
 from ._actions import _ActionConfigLoad
-from ._common import get_class_instantiator, get_generic_origin, is_dataclass_like, is_subclass
+from ._common import LoggerProperty, get_class_instantiator, get_generic_origin, is_dataclass_like, is_subclass
 from ._optionals import get_doc_short_description, is_pydantic_model, pydantic_support
 from ._parameter_resolvers import (
     ParamData,
@@ -21,7 +21,7 @@ from ._typehints import (
     callable_instances,
     is_optional,
 )
-from ._util import LoggerProperty, get_import_path, get_private_kwargs, iter_to_set_str
+from ._util import get_import_path, get_private_kwargs, iter_to_set_str
 from .typing import register_pydantic_type
 
 __all__ = [
